13-Aug-2020-Code.py
This is the updated code on 13 Aug 2020. M
Most of the code has been placed under a function so that the function can be called repeatedly and thus repeat the game.
This has opened up the option to ask the user if they would like to play again or exit the game.
Ideas:
   1. Create spacing so that text is easier to read
   2. Start learning how to build a GUI to make it more interactive.

Might still be buggy.